Use direct value on ng-pattern instead of function
The use of a method to get the ng-pattern value produces a recursive
call. This patch just uses a scope value directly in the ng-pattern
instead of a function.

This also reverts this commit: 6d7d16b86c25535208958459d30360d8e9653b1d
to bring back the regular expressions on patterns-service.