adding support for login with otp-code field (#233)
* adding support for login with otp-code field

* disabling user_id field for smart-link

Smart-link authentication method will be able to be done
by both the external application and Sequent IAM backend.

This means that Sequent will be able to verify the voter
authentication credentials. One of the credentials is the
user_id, which is a mandatory extra-field for smart-link
authnetication and comes from the auth-token received by
the external application. As such, it cannot be edited
by the user, and thus we are making it a filled-in and
disabled field in the login screen.

Future work will include allowing to hide a field and
showing it under a different name, to make mandatory
fields like user_id more user-friendly.

* fixing automatic request of otp code